+++ Source/AdmiralTheme/Classes/CoreFoundation/Themes/SwiftUI/AppThemeScheme.swift
@@ -1,0 +1,15 @@
+//
+//  AppThemeScheme.swift
+//  AdmiralTheme
+//
+//  Created by ADyatkov on 06.09.2022.
+//
+
+import Combine
+import SwiftUI
+
+@available(iOS 14.0.0, *)
+/// App theme scheme interface.
+public protocol AppThemeScheme {
+    init(theme: AppTheme)
+}
