+++ Source/AdmiralTheme/Classes/CoreFoundation/Themes/SwiftUI/SchemeProvider.swift
@@ -8,46 +8,28 @@
 import Combine
 import SwiftUI
 
-@available(iOS 14.0.0, *)
-public protocol AppThemeScheme {
-    init(theme: AppTheme)
-}
-
 @available (iOS 14.0.0, *)
-public class SchemeProvider<S>: ObservableObject {
-    @Published public fileprivate(set) var scheme: S
-    public init (scheme: S) {
-        self.scheme = scheme
-    }
-}
+/// Scheme provider is classs to store scheme.
+public class SchemeProvider<S>: ObservableObject where S: AppThemeScheme {
 
-@available (iOS 14.0.0, *)
-public class ManualSchemeProvider<S>: SchemeProvider<S> {
-    public func update(scheme: S) {
-        self.scheme = scheme
-    }
-}
+    // MARK: - Public Properties
 
-@available(iOS 14.0.0, *)
-public class AppThemeSchemeProvider<S>: SchemeProvider<S> where S: AppThemeScheme {
-    private let manager: SwiftUIThemeManager
-    private var subscribers: Set<AnyCancellable> = []
-    
-    public init(manager: SwiftUIThemeManager = .shared) {
-        self.manager = manager
-        super.init(scheme: S(theme: manager.theme))
-        manager.$theme.sink { [weak self] theme in
-            self?.scheme = S(theme: theme)
-        }.store(in: &subscribers)
-    }
-}
+    /// Scheme element. It store colors and fonts details.
+    @Published public internal(set) var scheme: S
 
-@available(iOS 14.0.0, *)
-public final class AppThemeConstantSchemeProvider<S> where S: AppThemeScheme {
-    public var scheme: S
+    // MARK: - Public Methods
+
+    /// Get scheme provider.
+    /// - Parameter scheme: Scheme element. It store colors and fonts details.
+    /// - Returns: Scheme provider.
+    public static func constant(scheme: S) -> SchemeProvider<S> {
+        return SchemeProvider<S>(scheme: scheme)
+    }
+
+    // MARK: - Initializer
+
     public init(scheme: S) {
         self.scheme = scheme
     }
+
 }
-
-
